remove byte casting in test
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene4547@1433728 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/test/org/apache/lucene/index/TestNorms.java
@@ -106,7 +106,7 @@ public class TestNorms extends LuceneTestCase {
     for (int i = 0; i < open.maxDoc(); i++) {
       StoredDocument document = open.document(i);
       int expected = Integer.parseInt(document.get(byteTestField));
-      assertEquals((byte)expected, (byte)normValues.get(i));
+      assertEquals(expected, normValues.get(i));
     }
     open.close();
     dir.close();
